+++ example/hello/hello.dart
@@ -8,32 +8,51 @@ import 'dart:html' as dart_html;
 import 'package:dart_web_toolkit/event.dart' as event;
 import 'package:dart_web_toolkit/shared.dart' as shared;
 import 'package:dart_web_toolkit/ui.dart' as ui;
+import 'package:dart_web_toolkit/util.dart' as util;
 
 void main() {
+  //*****************
+  // SplitLayoutPanel
+  //*****************
+
+  ui.SplitLayoutPanel p = new ui.SplitLayoutPanel();
+  p.addWest(new ui.Button("navigation"), 128.0);
+  p.addNorth(new ui.Button("list"), 384.0);
+  p.add(new ui.Button("details"));
+
+  // Attach the LayoutPanel to the RootLayoutPanel. The latter will listen for
+  // resize events on the window to ensure that its children are informed of
+  // possible size changes.
+  ui.RootLayoutPanel rp = ui.RootLayoutPanel.get();
+  rp.add(p);
+}
+
+void main2() {
   //****************
   // DockLayoutPanel
   //****************
-  
+
   ui.DockLayoutPanel p = new ui.DockLayoutPanel(util.Unit.PX);
   p.getElement().id = "dock";
   //
-  ui.Button btn = new ui.Button("north");
-  btn.getElement().id = "btn";
-  p.addNorth(btn, 200.0);
+  p.addNorth(new ui.Button("north"), 50.0);
+  p.addSouth(new ui.Button("south"), 100.0);
+  p.addEast(new ui.Button("east"), 150.0);
+  p.addWest(new ui.Button("west"), 200.0);
+  p.add(new ui.Button("center"));
+
   //
   ui.RootLayoutPanel root = ui.RootLayoutPanel.get();
   root.getElement().id = "root";
   root.add(p);
-  
-//  p.forceLayout();
-//  p.addSouth(new ui.Button("south"), 200.0);
-//  p.addEast(new ui.Button("east"), 200.0);
-//  p.addWest(new ui.Button("west"), 200.0);
-//  p.add(new ui.Button("center"));
 
 }
 
-void main2() {
+void main1() {
+  //*******
+  // Anchor
+  //*******
+
   ui.Anchor anchor = new ui.Anchor(true);
   print("Href: ${anchor.href}");
   //
@@ -71,41 +90,52 @@ void main2() {
 
   ui.RootPanel.get("testId").add(anchor);
 
+  //*******
+  // Button
+  //*******
+
+  ui.Button button = new ui.Button("Click me");
+  ui.RootPanel.get("testId").add(button);
+
+  //************
+  // HeaderPanel
+  //************
+
   ui.HeaderPanel headerPanel = new ui.HeaderPanel();
+
   headerPanel.setSize("150px", "100px");
 
   headerPanel.setHeaderWidget(new ui.Button("Header"));
   headerPanel.setFooterWidget(new ui.CheckBox("Active Footer"));
-  headerPanel.setContentWidget(new ui.Button("Center"));
+  headerPanel.setContentWidget(new ui.RadioButton("group", "Center"));
+
   ui.RootPanel.get("testId").add(headerPanel);
 
   //**************
   // VerticalPanel
   //**************
-  
+
   ui.VerticalPanel vPanel = new ui.VerticalPanel();
   vPanel.spacing = 5;
-  
+
   // Add some content to the panel
   for (int i = 1; i < 10; i++) {
     vPanel.add(new ui.Button("Button $i"));
   }
-  
+
   ui.RootPanel.get("testId").add(vPanel);
-  
+
   //**************
   // HorizontalPanel
   //**************
-  
+
   ui.HorizontalPanel hPanel = new ui.HorizontalPanel();
   hPanel.spacing = 5;
-  
+
   // Add some content to the panel
   for (int i = 1; i < 10; i++) {
     hPanel.add(new ui.Button("Button ${i}1"));
   }
-  
+
   ui.RootPanel.get("testId").add(hPanel);
-
-
 }