+++ lib/src/ui/dock_layout_panel.dart
@@ -6,25 +6,25 @@ part of dart_web_toolkit_ui;
 /**
  * A panel that lays its child widgets out "docked" at its outer edges, and
  * allows its last widget to take up the remaining space in its center.
- * 
+ *
  * <p>
  * This widget will <em>only</em> work in standards mode, which requires that
  * the HTML page in which it is run have an explicit &lt;!DOCTYPE&gt;
  * declaration.
  * </p>
- * 
+ *
  * <p>
  * <h3>Example</h3>
  * {@example com.google.gwt.examples.DockLayoutPanelExample}
  * </p>
- * 
+ *
  * <h3>Use in UiBinder Templates</h3>
  * <p>
- * DockLayoutPanel elements in  
+ * DockLayoutPanel elements in
  * {@link com.google.gwt.uibinder.client.UiBinder UiBinder} templates
  * lay out their children in elements tagged with the cardinal directions,
  * and center:
- * 
+ *
  * <p>
  * <dl>
  * <dt>&lt;g:center>
@@ -33,14 +33,14 @@ part of dart_web_toolkit_ui;
  * <dt>&lt;g:west>
  * <dt>&lt;g:east>
  * </dl>
- * 
+ *
  * <p>
  * Each child can hold only widget, and there can be only one &lt;g:center>.
  * However, there can be any number of the directional children.
  *<p>
  * (Note that the tags of the child elements are not
- * capitalized. This is meant to signal that they are not runtime objects, 
- * and so cannot have a <code>ui:field</code> attribute.) 
+ * capitalized. This is meant to signal that they are not runtime objects,
+ * and so cannot have a <code>ui:field</code> attribute.)
  * <p>
  * For example:<pre>
  * &lt;g:DockLayoutPanel unit='EM'>
@@ -69,7 +69,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   LayoutCommand _layoutCmd;
   double _filledWidth = 0.0;
   double _filledHeigh = 0.0;
-  
+
   /**
    * Creates an empty dock panel.
    *
@@ -82,11 +82,11 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     _layout = new Layout(getElement());
     _layoutCmd = new DockAnimateCommand(this, _layout);
   }
-  
+
   //*********
   // Children
   //*********
-  
+
   /**
    * Adds a widget at the center of the dock. No further widgets may be added
    * after this one.
@@ -96,9 +96,9 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void add(Widget widget) {
     _insert(widget, Direction.CENTER, 0.0, null);
   }
-  
+
   //*****
-  
+
   /**
    * Adds a widget to the east edge of the dock.
    *
@@ -108,16 +108,16 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addEast(Widget widget, double size) {
     _insert(widget, Direction.EAST, size, null);
   }
-  
+
   /**
    * Overloaded version for IsWidget.
-   * 
+   *
    * @see #addEast(Widget,double)
    */
   void addIsWidgetEast(IsWidget widget, double size) {
     this.addEast(widget.asWidget(), size);
   }
-  
+
   /**
    * Adds a widget to the end of the line. In LTR mode, the widget is added to
    * the east. In RTL mode, the widget is added to the west.
@@ -128,7 +128,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addLineEnd(Widget widget, double size) {
     _insert(widget, Direction.LINE_END, size, null);
   }
-  
+
   /**
    * Adds a widget to the start of the line. In LTR mode, the widget is added to
    * the west. In RTL mode, the widget is added to the east.
@@ -139,7 +139,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addLineStart(Widget widget, double size) {
     _insert(widget, Direction.LINE_START, size, null);
   }
-  
+
   /**
    * Adds a widget to the north edge of the dock.
    *
@@ -149,16 +149,16 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addNorth(Widget widget, double size) {
     _insert(widget, Direction.NORTH, size, null);
   }
-  
+
   /**
    * Overloaded version for IsWidget.
-   * 
+   *
    * @see #addNorth(Widget,double)
    */
   void addIsWidgetNorth(IsWidget widget, double size) {
     this.addNorth(widget.asWidget(), size);
   }
-  
+
   /**
    * Adds a widget to the south edge of the dock.
    *
@@ -168,16 +168,16 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addSouth(Widget widget, double size) {
     _insert(widget, Direction.SOUTH, size, null);
   }
-  
+
   /**
    * Overloaded version for IsWidget.
-   * 
+   *
    * @see #addSouth(Widget,double)
    */
   void addIsWidgetSouth(IsWidget widget, double size) {
     this.addSouth(widget.asWidget(), size);
   }
-  
+
   /**
    * Adds a widget to the west edge of the dock.
    *
@@ -187,16 +187,16 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void addWest(Widget widget, double size) {
     _insert(widget, Direction.WEST, size, null);
   }
-  
+
   /**
    * Overloaded version for IsWidget.
-   * 
+   *
    * @see #addWest(Widget,double)
    */
   void addIsWidgetWest(IsWidget widget, double size) {
     this.addWest(widget.asWidget(), size);
   }
-  
+
   /**
    * Adds a widget to the east edge of the dock, inserting it before an existing
    * widget.
@@ -209,7 +209,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertEast(Widget widget, double size, Widget before) {
     _insert(widget, Direction.EAST, size, before);
   }
-  
+
   /**
    * Adds a widget to the start of the line, inserting it before an existing
    * widget. In LTR mode, the widget is added to the east. In RTL mode, the
@@ -223,7 +223,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertLineEnd(Widget widget, double size, Widget before) {
     _insert(widget, Direction.LINE_END, size, before);
   }
-  
+
   /**
    * Adds a widget to the end of the line, inserting it before an existing
    * widget. In LTR mode, the widget is added to the west. In RTL mode, the
@@ -237,7 +237,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertLineStart(Widget widget, double size, Widget before) {
     _insert(widget, Direction.LINE_START, size, before);
   }
-  
+
   /**
    * Adds a widget to the north edge of the dock, inserting it before an
    * existing widget.
@@ -250,7 +250,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertNorth(Widget widget, double size, Widget before) {
     _insert(widget, Direction.NORTH, size, before);
   }
-  
+
   /**
    * Adds a widget to the south edge of the dock, inserting it before an
    * existing widget.
@@ -263,7 +263,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertSouth(Widget widget, double size, Widget before) {
     _insert(widget, Direction.SOUTH, size, before);
   }
-  
+
   /**
    * Adds a widget to the west edge of the dock, inserting it before an existing
    * widget.
@@ -276,7 +276,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   void insertWest(Widget widget, double size, Widget before) {
     _insert(widget, Direction.WEST, size, before);
   }
-  
+
   /**
    * Adds a widget to the specified edge of the dock. If the widget is already a
    * child of this panel, this method behaves as though {@link #remove(Widget)}
@@ -287,7 +287,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param before the widget before which to insert the new child, or
    *          <code>null</code> to append
    */
-  void _insert(Widget widget, int direction, double size, Widget before) {
+  void _insert(Widget widget, Direction direction, double size, Widget before) {
     assertIsChild(before);
 
     // Validation.
@@ -324,26 +324,26 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     // Update the layout.
     animate(0);
   }
-  
+
   //**********
   // Animation
   //**********
-  
+
   void animate(int duration, [LayoutAnimationCallback callback = null]) {
     _layoutCmd.schedule(duration, callback);
   }
-  
+
   void forceLayout() {
     _layoutCmd.cancel();
     _doLayout();
     _layout.layout();
     onResize();
   }
-  
+
   //***********
   // LayoutData
   //***********
-  
+
   /**
    * Gets the container element wrapping the given child widget.
    *
@@ -354,7 +354,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     assertIsChild(child);
     return (child.getLayoutData() as LayoutData).layer.getContainerElement();
   }
-  
+
   /**
    * Gets the layout direction of the given child widget.
    *
@@ -363,17 +363,17 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *         child of this panel
    * @throws AssertionError if the widget is not a child and assertions are enabled
    */
-  int getWidgetDirection(Widget child) {
+  Direction getWidgetDirection(Widget child) {
     assertIsChild(child);
     if (child.getParent() != this) {
       return null;
     }
     return (child.getLayoutData() as LayoutData).direction;
   }
-  
+
   /**
    * Gets the layout size of the given child widget.
-   * 
+   *
    * @param child the widget to be queried
    * @return the widget's layout size, or <code>null</code> if it is not a child of
    *         this panel
@@ -386,9 +386,9 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     }
     return (child.getLayoutData() as LayoutData).size;
   }
-  
+
   //***********
-  
+
   void onResize() {
     for (Widget child in getChildren()) {
       if (child is RequiresResize) {
@@ -396,7 +396,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
       }
     }
   }
-  
+
   bool remove(Widget w) {
     bool removed = super.remove(w);
     if (removed) {
@@ -411,27 +411,27 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
 
     return removed;
   }
-  
+
   //************
-  
+
   /**
    * Sets whether or not the given widget should be hidden.
-   * 
+   *
    * @param widget the widget to hide or display
    * @param hidden true to hide the widget, false to display it
    */
   void setWidgetHidden(Widget widget, bool hidden) {
     assertIsChild(widget);
-    
+
     LayoutData data = widget.getLayoutData() as LayoutData;
     if (data.hidden == hidden) {
       return;
     }
-    
+
     data.hidden = hidden;
     animate(0);
   }
-  
+
   /**
    * Updates the size of the widget passed in as long as it is not the center
    * widget and updates the layout of the dock.
@@ -450,7 +450,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     // Update the layout.
     animate(0);
   }
-  
+
   Widget getCenter() {
     return _center;
   }
@@ -462,7 +462,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
   double getCenterWidth() {
     return getElement().clientWidth / _layout.getUnitSize(_unit, false) - _filledWidth;
   }
-  
+
   /**
    * Resolve the specified direction based on the current locale. If the
    * direction is {@link Direction#LINE_START} or {@link Direction#LINE_END},
@@ -473,7 +473,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param direction the specified direction
    * @return the locale
    */
-  int getResolvedDirection(int direction) {
+  Direction getResolvedDirection(Direction direction) {
 //    if (direction == Direction.LINE_START) {
 //      return LocaleInfo.getCurrentLocale().isRTL()
 //          ? Direction.EAST : Direction.WEST;
@@ -483,31 +483,31 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
 //    }
     return direction;
   }
-  
+
   Unit getUnit() {
     return _unit;
   }
-  
+
   //************
-  
+
   void onAttach() {
     super.onAttach();
     _layout.onAttach();
   }
-  
+
   void onDetach() {
     super.onDetach();
     _layout.onDetach();
   }
-  
+
   void assertIsChild(Widget widget) {
     assert ((widget == null) || (widget.getParent() == this)); // : "The specified widget is not a child of this panel";
   }
-  
+
   //*******
   // Layout
   //*******
-  
+
   void _doLayout() {
     double left = 0.0;
     double top = 0.0;
@@ -567,27 +567,30 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
  * Used in {@link DockLayoutPanel#addEast(Widget, double)} et al to specify
  * the direction in which a child widget will be added.
  */
-class Direction {
-  static const int NORTH = 0;
-  static const int EAST = 1;
-  static const int SOUTH = 2;
-  static const int WEST = 3;
-  static const int CENTER = 4;
-  static const int LINE_START = 5;
-  static const int LINE_END = 6;
+class Direction<int> extends Enum<int> {
+
+  const Direction(int type) : super (type);
+
+  static const Direction NORTH = const Direction(0);
+  static const Direction EAST = const Direction(1);
+  static const Direction SOUTH = const Direction(2);
+  static const Direction WEST = const Direction(3);
+  static const Direction CENTER = const Direction(4);
+  static const Direction LINE_START = const Direction(5);
+  static const Direction LINE_END = const Direction(6);
 }
 
 /**
  * Layout data associated with each widget.
  */
 class LayoutData {
-  int direction;
+  Direction direction;
   double oldSize, size;
   double originalSize;
   bool hidden = false;
   Layer layer;
 
-  LayoutData(int direction, double size, Layer layer) {
+  LayoutData(Direction direction, double size, Layer layer) {
     this.direction = direction;
     this.size = size;
     this.layer = layer;
@@ -595,9 +598,9 @@ class LayoutData {
 }
 
 class DockAnimateCommand extends LayoutCommand {
-  
+
   DockLayoutPanel _panel;
-  
+
   DockAnimateCommand(this._panel, Layout layout) : super(layout);
 
   void doBeforeLayout() {